+++ inferno-ml-server-types/src/Inferno/ML/Server/Types.hs
@@ -17,7 +17,7 @@ import Conduit (ConduitT)
 import Control.Applicative (asum, optional)
 import Control.Category ((>>>))
 import Control.DeepSeq (NFData (rnf), rwhnf)
-import Control.Monad (void)
+import Control.Monad (void, (<=<))
 import Data.Aeson
 import Data.Aeson.Types (Parser)
 import qualified Data.Attoparsec.ByteString.Char8 as Attoparsec
@@ -64,6 +64,7 @@ import Database.PostgreSQL.Simple.Types
   )
 import Foreign.C (CUInt (CUInt))
 import GHC.Generics (Generic)
+import Inferno.Types.Syntax (Ident)
 import Inferno.Types.VersionControl
   ( VCObjectHash,
     byteStringToVCObjectHash,
@@ -172,8 +173,9 @@ newtype Id a = Id Int64
 
 -- | Row for the table containing inference script closures
 data InferenceScript uid gid = InferenceScript
-  { -- NOTE: This is the ID for each row, stored as a `bytea` (bytes of the hash)
+  { -- | This is the ID for each row, stored as a @bytea@ (bytes of the hash)
     hash :: VCObjectHash,
+    -- | Script closure
     obj :: VCMeta uid gid VCObject
   }
   deriving stock (Show, Eq, Generic)
@@ -237,6 +239,8 @@ data Model uid gid = Model
     -- | The user who owns the model, if any. Note that owning a model
     -- will implicitly set permissions
     user :: Maybe uid,
+    -- | The time that this model was \"deleted\", if any. For active models,
+    -- this will be @Nothing@
     terminated :: Maybe UTCTime
   }
   deriving stock (Show, Eq, Generic)
@@ -341,6 +345,8 @@ data ModelVersion uid gid c = ModelVersion
     -- the PSQL large object table
     contents :: c,
     version :: Version,
+    -- | The time that this model version was \"deleted\", if any. For active
+    -- models versions, this will be @Nothing@
     terminated :: Maybe UTCTime
   }
   deriving stock (Show, Eq, Generic)
@@ -592,13 +598,24 @@ data InferenceParam uid gid p s = InferenceParam
     -- (e.g. a UUID for use with @inferno-lsp@)
     --
     -- For existing inference params, this is the foreign key for the specific
-    -- script in the 'InferenceScript' table
+    -- script in the 'InferenceScript' table (i.e. a @VCObjectHash@)
     script :: s,
     -- | This needs to be linked to a specific version of a model rather
     -- than the @model@ table itself
     model :: Id (ModelVersion uid gid Oid),
-    inputs :: Vector (SingleOrMany p),
-    outputs :: Vector (SingleOrMany p),
+    -- | This is called @inputs@ but is also used for script outputs as
+    -- well. The access (input or output) is controlled by the 'ScriptInputType'.
+    -- For example, if this field is set to @[("input0", Single (p, Readable))]@,
+    -- the script will only have a single read-only input and will not be able to
+    -- write anywhere (note that we should disallow this scenario, as script
+    -- evaluation would not work properly)
+    --
+    -- Mapping the input\/output to the Inferno identifier helps ensure that
+    -- Inferno identifiers are always pointing to the correct input\/output;
+    -- otherwise we would need to rely on the order of the original identifiers
+    inputs :: Map Ident (SingleOrMany p, ScriptInputType),
+    -- | The time that this parameter was \"deleted\", if any. For active
+    -- parameters, this will be @Nothing@
     terminated :: Maybe UTCTime,
     user :: uid
   }
@@ -619,10 +636,6 @@ instance
       <$> field
       <*> fmap wrappedTo (field @VCObjectHashRow)
       <*> field
-      -- HACK / FIXME This is a pretty awful hack (storing as `jsonb`),
-      -- but Postgres sub-arrays need to be the same length and writing
-      -- a custom parser might be painful
-      <*> fmap getAeson field
       <*> fmap getAeson field
       <*> field
       <*> field
@@ -638,12 +651,40 @@ instance
     [ toField Default,
       ip ^. the @"script" & VCObjectHashRow & toField,
       ip ^. the @"model" & toField,
-      -- HACK / FIXME See above
       ip ^. the @"inputs" & Aeson & toField,
-      ip ^. the @"outputs" & Aeson & toField,
       toField Default,
       ip ^. the @"user" & toField
     ]
+
+-- | Controls input interaction within a script, i.e. ability to read from
+-- and\/or write to this input. Although the term \"input\" is used, those with
+-- writes enabled can also be described as \"outputs\"
+data ScriptInputType
+  = -- | Script input can be read, but not written
+    Readable
+  | -- | Script input can be written, i.e. can be used in array of
+    -- write objects returned from script evaluation
+    Writable
+  | -- | Script input can be both read from and written to; this allows
+    -- the same script identifier to point to the same PID with both
+    -- types of access enabled
+    ReadableWritable
+  deriving stock (Show, Eq, Generic)
+  deriving anyclass (NFData)
+
+instance FromJSON ScriptInputType where
+  parseJSON = withText "ScriptInputType" $ \case
+    "r" -> pure Readable
+    "w" -> pure Writable
+    "rw" -> pure ReadableWritable
+    s -> fail $ "Invalid script input type: " <> Text.unpack s
+
+instance ToJSON ScriptInputType where
+  toJSON =
+    String . \case
+      Readable -> "r"
+      Writable -> "w"
+      ReadableWritable -> "rw"
 
 -- | Information about execution time and resource usage. This is saved by
 -- @inferno-ml-server@ after script evaluation completes and can be queried
@@ -762,22 +803,42 @@ instance FromJSON IValue where
     Number n -> pure . IDouble $ toRealFloat n
     -- It's easier to just mark the time explicitly in an object,
     -- rather than try to deal with distinguishing times and doubles
-    Object o -> ITime <$> o .: "time"
+    Object o ->
+      asum
+        [ ITime <$> o .: "time",
+          fmap IArray $ arrayP =<< o .: "array"
+        ]
+    -- Note that this preserves a plain JSON array for tuples. But we need
+    -- some straightforward way of distinguishing tuples and arrays; since
+    -- the bridge often transmits a large number of individual tuples (times
+    -- and values), it's better to use arrays for the tuples and a tagged object
+    -- for arrays themselves; we often will only deal with one large array, and
+    -- adding a few bytes to this is better than adding a few bytes to thousands
+    -- of encoded tuples
     Array a
       | [x, y] <- Vector.toList a ->
-          (,) <$> parseJSON x <*> parseJSON y <&> \case
-            -- We don't want to confuse a two-element array of tuples with
-            -- a tuple itself. For example, `"[[10.0, {\"time\": 10}], [10.0, {\"time\": 10}]]"`
-            -- should parse as a two-element array of `(double, time)` tuples,
-            -- not as a `((double, time), (double, time))`. I can't think of
-            -- any reason to support the latter. An alternative would be to
-            -- change tuple encoding to an object, but then we would be transmitting
-            -- a much larger amount of data on most requests
-            (f@(ITuple _), s@(ITuple _)) -> IArray $ Vector.fromList [f, s]
-            t -> ITuple t
-      | otherwise -> IArray <$> traverse (parseJSON @IValue) a
+          fmap ITuple $ (,) <$> parseJSON x <*> parseJSON y
+      | otherwise -> fail "Only two-element tuples are supported"
     Null -> pure IEmpty
-    _ -> fail "Expected one of: string, double, time, tuple, unit (empty array), array"
+    _ -> fail "Expected one of: string, double, time, tuple, null, array"
+    where
+      arrayP :: Vector Value -> Parser (Vector IValue)
+      arrayP a =
+        -- This is a bit tedious, but we want to make sure that the array elements
+        -- are homogeneous; parsing all elements to `IValue`s first can't guarantee
+        -- this
+        asum
+          [ -- This alternative means that `null` will be correctly parsed to NaN
+            -- when inside an array of doubles
+            fmap IDouble <$> traverse parseJSON a,
+            fmap ITuple <$> traverse parseJSON a,
+            fmap IText <$> traverse parseJSON a,
+            fmap ITime <$> traverse (withObject "EpochTime" (.: "time")) a,
+            -- Nested array support
+            fmap IArray
+              <$> traverse (withObject "IArray" (arrayP <=< (.: "array"))) a,
+            fail "Expected a heterogeneous array"
+          ]
 
 instance ToJSON IValue where
   toJSON = \case
@@ -786,8 +847,9 @@ instance ToJSON IValue where
     ITuple t -> toJSON t
     -- See `FromJSON` instance above
     ITime t -> object ["time" .= t]
+    -- See `FromJSON` instance above
+    IArray is -> object ["array" .= is]
     IEmpty -> toJSON Null
-    IArray is -> toJSON is
 
 -- | Used to represent inputs to the script. 'Many' allows for an array input
 data SingleOrMany a