+++ inferno-ml-server/src/Inferno/ML/Server/Inference.hs
@@ -31,7 +31,6 @@ import Data.Time (UTCTime, getCurrentTime)
 import Data.Time.Clock.POSIX (getPOSIXTime)
 import Data.Traversable (for)
 import Data.UUID (UUID)
-import qualified Data.Vector as Vector
 import Data.Word (Word64)
 import Database.PostgreSQL.Simple
   ( Only (Only),
@@ -92,12 +91,14 @@ import UnliftIO.IORef (readIORef)
 import UnliftIO.MVar (putMVar, takeMVar, withMVar)
 import UnliftIO.Timeout (timeout)
 
--- Run an inference param, locking the `MVar` held in the `Env`. This is to avoid
--- running params in parallel (which may lead to problems with model caching,
--- etc...) and also to indicate whether any active process is running in the
--- `status` endpoint (using `tryTakeMVar`)
+-- | Run an inference param, locking the @MVar@ held in the 'Env'. This is to
+-- avoid running params in parallel (which may lead to problems with model
+-- caching, etc...) and also to indicate whether any active process is running
+-- in the @/status@ endpoint (using @tryTakeMVar@)
 runInferenceParam ::
   Id InferenceParam ->
+  -- | Optional resolution, defaulting to 128. This is needed in case the
+  -- parameter evaluates a script that calls e.g. @valueAt@
   Maybe Int64 ->
   UUID ->
   RemoteM (WriteStream IO)
@@ -167,9 +168,17 @@ runInferenceParam ipid (fromMaybe 128 -> res) uuid =
                               ( mkIdent i,
                                 pids ^.. each & over mapped toSeries & VArray
                               )
-                        where
-                          ps :: [SingleOrMany PID]
-                          ps = param ^.. #inputs . each
+
+                      -- Note that this both includes inputs (i.e. readable)
+                      -- and outputs (i.e. writable, or readable/writable).
+                      -- These need to be provided to the script in order
+                      -- for the symbolic identifer (e.g. `output0`) to
+                      -- resolve. We can discard the input type here,
+                      -- however. The distinction is only relevant for the
+                      -- runtime that runs as a script evaluation engine
+                      -- and commits the output write object
+                      ps :: [SingleOrMany PID]
+                      ps = param ^.. #inputs . to Map.toAscList . each . _2 . _1
 
                       closure :: Map VCObjectHash VCObject
                       closure =
@@ -185,9 +194,10 @@ runInferenceParam ipid (fromMaybe 128 -> res) uuid =
                           )
                           args
                         where
+                          -- See note above about inputs/outputs
                           args :: [Expr (Maybe a) ()]
                           args =
-                            [0 .. param ^. #inputs & Vector.length & (- 1)]
+                            [0 .. length ps - 1]
                               <&> Var () Nothing LocalScope
                                 . Expl
                                 . ExtIdent
@@ -198,18 +208,10 @@ runInferenceParam ipid (fromMaybe 128 -> res) uuid =
                           dummy :: ImplExpl
                           dummy = Expl . ExtIdent $ Right "dummy"
 
-                  doEval expr =<< runImplEnvM mempty (mkEnvFromClosure localEnv closure)
+                  either (throwInfernoError . Left . SomeInfernoError) yieldPairs
+                    =<< flip (`evalExpr` implEnv) expr
+                    =<< runImplEnvM mempty (mkEnvFromClosure localEnv closure)
                   where
-                    doEval ::
-                      Expr (Maybe VCObjectHash) () ->
-                      BridgeTermEnv RemoteM ->
-                      RemoteM (WriteStream IO)
-                    doEval x env =
-                      either
-                        (throwInfernoError . Left . SomeInfernoError)
-                        yieldPairs
-                        =<< evalExpr env implEnv x
-
                     yieldPairs ::
                       Value BridgeMlValue (ImplEnvM RemoteM BridgeMlValue) ->
                       RemoteM (WriteStream IO)