+++ inferno-ml-server/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 2023.5.29
+* Change representation of script inputs/outputs
+
 ## 2023.5.22
 * Add support for tracking evaluation info